+++ packages/core/src/index.ts
@@ -1,4 +1,4 @@
-export const VERSION = "0.8.0";
+export const VERSION = "0.8.1";
 
 export type {
 	CreateComponentArgs,